# api/src/services/swagger_output.json
{
  "openapi": "3.1.1",
  "info": {
    "version": "1.0.0",
    "title": "Bingo Player API",
    "description": "Documentation of the Bingo Player API"
  },
  "servers": [
    {
      "url": "http://localhost:8080/"
    }
  ],
  "paths": {
...
    "/api/v1/spaces/get/all/list/{prop}": {
      "get": {
        "tags": [
          "Spaces",
          "List"
        ],
        "description": "  Get all values for a given space query prop.  The values are returned as an array of strings, but may contain null values.  ",
        "parameters": [
          {
            "name": "prop",
            "in": "path",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/SpaceQueryProp"
            },
            "description": "The space query prop to list.",
            "type": "string"
          },
          {
            "$ref": "#/components/parameters/offset"
          },
          {
            "$ref": "#/components/parameters/o"
          },
          {
            "$ref": "#/components/parameters/limit"
          },
          {
            "$ref": "#/components/parameters/l"
          },
          {
            "$ref": "#/components/parameters/sortOrder"
          },
          {
            "$ref": "#/components/parameters/so"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListSpaceProp200ResponseBody"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/ListSpaceProp200ResponseBody"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse400"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse400"
                }
              }
            }
          },
          "404": {
            "description": "No matching spaces found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse404"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse404"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "501": {
            "description": "Not implemented.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse501"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse501"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/spaces/search/{query}": {
      "get": {
        "tags": [
          "Spaces",
          "Get",
          "Search"
        ],
        "description": "Use the Atlas Search index to search for spaces based on the provided query string.",
        "parameters": [
          {
            "name": "query",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "The search query."
          },
          {
            "$ref": "#/components/parameters/limit"
          },
          {
            "$ref": "#/components/parameters/l"
          },
          {
            "$ref": "#/components/parameters/offset"
          },
          {
            "$ref": "#/components/parameters/o"
          },
          {
            "$ref": "#/components/parameters/sortByWithScore"
          },
          {
            "$ref": "#/components/parameters/sbWithScore"
          },
          {
            "$ref": "#/components/parameters/sortOrderDescDefault"
          },
          {
            "$ref": "#/components/parameters/soDescDefault"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SearchSpaces200ResponseBody"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/SearchSpaces200ResponseBody"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse400"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse400"
                }
              }
            }
          },
          "404": {
            "description": "No matching spaces found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse404"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse404"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "501": {
            "description": "Not implemented.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse501"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse501"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/spaces/search/{query}/list/{prop}": {
      "get": {
        "tags": [
          "Spaces",
          "List",
          "Search"
        ],
        "description": "  Use the Atlas Search index to get all values for a given space query prop that appear in the spaces that match  the given query. The values are returned as an array of objects, each containing a value and a relevance score  for the document containing that value with the highest relevance. The value is a string, but may be null.  ",
        "parameters": [
          {
            "name": "query",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "The search query."
          },
          {
            "name": "prop",
            "in": "path",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/SpaceQueryProp"
            },
            "description": "The space query prop to list.",
            "type": "string"
          },
          {
            "$ref": "#/components/parameters/limit"
          },
          {
            "$ref": "#/components/parameters/l"
          },
          {
            "$ref": "#/components/parameters/offset"
          },
          {
            "$ref": "#/components/parameters/o"
          },
          {
            "$ref": "#/components/parameters/sortOrderDescDefault"
          },
          {
            "$ref": "#/components/parameters/soDescDefault"
          },
          {
            "$ref": "#/components/parameters/sortBySearchListProp"
          },
          {
            "$ref": "#/components/parameters/sbSearchListProp"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SearchSpacesListProp200ResponseBody"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/SearchSpacesListProp200ResponseBody"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse400"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse400"
                }
              }
            }
          },
          "404": {
            "description": "No matching spaces found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse404"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse404"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "501": {
            "description": "Not implemented.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse501"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse501"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "parameters": {
      "limit": {
        "schema": {
          "$ref": "#/components/schemas/limit"
        },
        "required": false,
        "name": "limit",
        "in": "query"
      },
      "l": {
        "schema": {
          "$ref": "#/components/schemas/l"
        },
        "required": false,
        "name": "l",
        "in": "query"
      },
      "offset": {
        "schema": {
          "$ref": "#/components/schemas/offset"
        },
        "required": false,
        "name": "offset",
        "in": "query"
      },
      "o": {
        "schema": {
          "$ref": "#/components/schemas/o"
        },
        "required": false,
        "name": "o",
        "in": "query"
      },
...
      "sortOrder": {
        "schema": {
          "$ref": "#/components/schemas/sortOrder"
        },
        "required": false,
        "name": "sortOrder",
        "in": "query"
      },
      "so": {
        "schema": {
          "$ref": "#/components/schemas/so"
        },
        "required": false,
        "name": "so",
        "in": "query"
      },
      "sortOrderDescDefault": {
        "schema": {
          "$ref": "#/components/schemas/sortOrderDescDefault"
        },
        "required": false,
        "name": "sortOrder",
        "in": "query"
      },
      "soDescDefault": {
        "schema": {
          "$ref": "#/components/schemas/soDescDefault"
        },
        "required": false,
        "name": "so",
        "in": "query"
      },
...
      "sortByWithScore": {
        "schema": {
          "$ref": "#/components/schemas/sortByWithScore"
        },
        "required": false,
        "name": "sortBy",
        "in": "query"
      },
      "sbWithScore": {
        "schema": {
          "$ref": "#/components/schemas/sbWithScore"
        },
        "required": false,
        "name": "sb",
        "in": "query"
      },
      "sortBySearchListProp": {
        "schema": {
          "$ref": "#/components/schemas/sortBySearchListProp"
        },
        "required": false,
        "name": "sortBy",
        "in": "query"
      },
      "sbSearchListProp": {
        "schema": {
          "$ref": "#/components/schemas/sbSearchListProp"
        },
        "required": false,
        "name": "sb",
        "in": "query"
      },
...
    },
    "schemas": {
...
      "SpaceQueryProp": {
        "type": "string",
        "enum": [
          "id",
          "name",
          "description",
          "examples",
          "aliases",
          "tags",
          "_id",
          "known-as",
          "alias",
          "tag",
          "example"
        ],
        "description": "A property of a space that can be queried."
      },
...
      "ErrorCode": {
        "type": "string",
        "enum": [
          "UNKNOWN",
          "BAD_INPUT",
          "BAD_URL",
          "BAD_QUERY",
          "BAD_BODY",
          "INTERNAL",
          "NOT_FOUND",
          "NOT_IMPLEMENTED",
          "VALIDATION"
        ],
        "description": "An error code that may be returned by the server."
      },
      "ErrorMessage": {
        "anyOf": [
          {
            "type": "string"
          },
          {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          {
            "type": "null"
          }
        ],
        "description": "A message describing an error.",
        "example": "Internal server error."
      },
      "HelpLink": {
        "type": "string",
        "format": "uri",
        "description": "A URL to the docs page that may help the user resolve the error or understand the response.",
        "example": "https://example.com/docs"
      },
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "ok": {
            "type": "boolean",
            "enum": [
              false
            ],
            "description": "Whether the operation was successful.",
            "example": false
          },
          "help": {
            "$ref": "#/components/schemas/HelpLink"
          },
          "code": {
            "allOf": [
              {
                "$ref": "#/components/schemas/ErrorCode"
              },
              {
                "description": "The error code.",
                "example": "UNKNOWN"
              }
            ]
          },
          "message": {
            "$ref": "#/components/schemas/ErrorMessage"
          }
        },
        "required": [
          "ok",
          "code",
          "message"
        ]
      },
      "ErrorResponse501": {
        "type": "object",
        "properties": {
          "ok": {
            "type": "boolean",
            "enum": [
              false
            ],
            "description": "Whether the operation was successful.",
            "example": false
          },
          "help": {
            "$ref": "#/components/schemas/HelpLink"
          },
          "code": {
            "type": "string",
            "enum": [
              "NOT_IMPLEMENTED"
            ],
            "description": "The error code."
          },
          "message": {
            "allOf": [
              {
                "$ref": "#/components/schemas/ErrorMessage"
              },
              {
                "example": "This feature is not yet implemented."
              }
            ]
          }
        },
        "required": [
          "ok",
          "code",
          "message"
        ]
      },
      "ErrorResponse400": {
        "type": "object",
        "properties": {
          "ok": {
            "type": "boolean",
            "enum": [
              false
            ],
            "description": "Whether the operation was successful.",
            "example": false
          },
          "help": {
            "$ref": "#/components/schemas/HelpLink"
          },
          "code": {
            "type": "string",
            "enum": [
              "BAD_INPUT",
              "BAD_BODY",
              "BAD_QUERY",
              "BAD_URL"
            ],
            "description": "The error code."
          },
          "message": {
            "allOf": [
              {
                "$ref": "#/components/schemas/ErrorMessage"
              },
              {
                "example": "Invalid input."
              }
            ]
          }
        },
        "required": [
          "ok",
          "code",
          "message"
        ]
      },
      "ErrorResponse404": {
        "type": "object",
        "properties": {
          "ok": {
            "type": "boolean",
            "enum": [
              false
            ],
            "description": "Whether the operation was successful.",
            "example": false
          },
          "help": {
            "$ref": "#/components/schemas/HelpLink"
          },
          "code": {
            "type": "string",
            "enum": [
              "NOT_FOUND"
            ],
            "description": "The error code."
          },
          "message": {
            "allOf": [
              {
                "$ref": "#/components/schemas/ErrorMessage"
              },
              {
                "example": "No resources found."
              }
            ]
          }
        },
        "required": [
          "ok",
          "code",
          "message"
        ]
      },
      "SuccessResponseBase": {
        "type": "object",
        "properties": {
          "ok": {
            "type": "boolean",
            "enum": [
              true
            ],
            "description": "Whether the operation was successful.",
            "example": true
          },
          "help": {
            "$ref": "#/components/schemas/HelpLink"
          }
        },
        "required": [
          "ok"
        ],
        "description": "A successful response from the server."
      },
...
      "limit": {
        "type": "integer",
        "minimum": 1,
        "description": "[Query Parameter] The maximum number of resources to return. Must be a positive integer and non-zero. If not provided, the server will return all found resources.",
        "example": 10
      },
      "l": {
        "type": "integer",
        "minimum": 1,
        "description": "[Query Parameter] Shorthand for limit.",
        "example": 10
      },
      "offset": {
        "type": [
          "integer",
          "null"
        ],
        "minimum": 0,
        "default": 0,
        "description": "[Query Parameter] The number of spaces to skip before returning results. Must be a non-negative integer. If not provided, it will be considered 0."
      },
      "o": {
        "type": [
          "integer",
          "null"
        ],
        "minimum": 0,
        "default": 0,
        "description": "[Query Parameter] Shorthand for offset."
      },
...
      "sortOrder": {
        "type": "string",
        "enum": [
          "asc",
          "ascending",
          "desc",
          "descending",
          "random",
          "rand"
        ],
        "default": "asc",
        "description": "[Query Parameter] The order to sort the field by. Ascending if not provided.",
        "example": "asc"
      },
      "so": {
        "type": "string",
        "enum": [
          "asc",
          "ascending",
          "desc",
          "descending",
          "random",
          "rand"
        ],
        "default": "asc",
        "description": "[Query Parameter] Shorthand for sortOrder.",
        "example": "asc"
      },
...
      "SearchSpaces200ResponseBody": {
        "allOf": [
          {
            "type": "object",
            "properties": {
              "spaces": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/CleanMongoSpace"
                },
                "description": "The spaces that match the query."
              }
            },
            "required": [
              "spaces"
            ],
            "description": "The response from searching spaces."
          },
          {
            "$ref": "#/components/schemas/SuccessResponseBase"
          }
        ]
      },
...
      "SearchSpacesListProp200ResponseBody": {
        "allOf": [
          {
            "type": "object",
            "properties": {
              "values": {
                "type": "array",
                "items": {
                  "type": "object",
                  "properties": {
                    "value": {
                      "anyOf": [
                        {
                          "type": "string"
                        },
                        {
                          "type": "null"
                        },
                        {
                          "type": "null"
                        }
                      ],
                      "description": "The value of the property."
                    },
                    "_score": {
                      "type": "number",
                      "description": "The relevance score of the value."
                    }
                  },
                  "required": [
                    "value",
                    "_score"
                  ]
                },
                "description": "The values found and their relevance scores."
              }
            },
            "required": [
              "values"
            ],
            "description": "List of values found"
          },
          {
            "$ref": "#/components/schemas/SuccessResponseBase"
          }
        ]
      },
...
      "ListSpaceProp200ResponseBody": {
        "allOf": [
          {
            "type": "object",
            "properties": {
              "values": {
                "type": "array",
                "items": {
                  "anyOf": [
                    {
                      "type": "string"
                    },
                    {
                      "type": "null"
                    },
                    {
                      "type": "null"
                    }
                  ]
                },
                "description": "The values of the property."
              }
            },
            "required": [
              "values"
            ],
            "description": "List of values found"
          },
          {
            "$ref": "#/components/schemas/SuccessResponseBase"
          }
        ]
      },
...
    }
  }
}